Create new tests_aio command
