+++ src/python/grpcio_tests/tests_aio/end2end/client_test.py
@@ -20,10 +20,10 @@ import unittest
 
 from grpc.experimental import aio
 from src.proto.grpc.testing import messages_pb2
-from tests.aio.end2end import sync_server
+from tests_aio.end2end import sync_server
 
 
-class TestChannel(unittest.TestCase):
+class TestClient(unittest.TestCase):
     def setUp(self):
         self._server = sync_server.Server()
         self._server.start()
@@ -32,15 +32,19 @@ class TestChannel(unittest.TestCase):
     def tearDown(self):
         self._server.terminate()
 
-    def test_async_context(self):
+    def test_unary_unary(self):
         async def coro():
-            async with aio.insecure_channel('localhost:%d' % sync_server.Server.PORT) as channel:
-                hi = channel.unary_unary(
-                    '/grpc.testing.TestService/UnaryCall',
-                    request_serializer=messages_pb2.SimpleRequest.SerializeToString,
-                    response_deserializer=messages_pb2.SimpleResponse.FromString
-                )
-                response = await hi(messages_pb2.SimpleRequest())
+            channel = aio.insecure_channel('localhost:%d' % sync_server.Server.PORT)
+            hi = channel.unary_unary(
+                '/grpc.testing.TestService/UnaryCall',
+                request_serializer=messages_pb2.SimpleRequest.SerializeToString,
+                response_deserializer=messages_pb2.SimpleResponse.FromString
+            )
+            response = await hi(messages_pb2.SimpleRequest())
+
+            self.assertEqual(type(response), messages_pb2.SimpleResponse)
+
+            await channel.close()
 
         asyncio.get_event_loop().run_until_complete(coro())
